requirements.txt
aiogram==3.20.0.post0
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
redis==6.0.0
celery==5.5.2
httpx==0.28.1
python-dotenv==1.1.0
orjson==3.10.18
alembic==1.15.2